backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.31.0
alembic==1.14.1
pydantic==2.10.5
pydantic-settings==2.7.1
python-multipart==0.0.20
pytesseract==0.3.13
pdf2image==1.17.0
anthropic==0.43.0
google-genai>=1.0.0
httpx==0.28.1
python-dotenv==1.0.1
sentry-sdk[fastapi]==2.19.2
supabase==2.11.0
PyJWT==2.10.1
cryptography==44.0.0
Pillow==11.1.0
pdfplumber==0.11.9
